factorizing code
