+++ src/main/java/com/github/syr0ws/fastinventory/common/placeholder/CommonPlaceholder.java
@@ -5,6 +5,8 @@ import com.github.syr0ws.fastinventory.common.placeholder.inventory.InventorySiz
 import com.github.syr0ws.fastinventory.common.placeholder.inventory.InventoryTypePlaceholder;
 import com.github.syr0ws.fastinventory.common.placeholder.item.ItemSlotPlaceholder;
 import com.github.syr0ws.fastinventory.common.placeholder.pagination.CurrentPagePlaceholder;
+import com.github.syr0ws.fastinventory.common.placeholder.pagination.NextPagePlaceholder;
+import com.github.syr0ws.fastinventory.common.placeholder.pagination.PreviousPagePlaceholder;
 import com.github.syr0ws.fastinventory.common.placeholder.pagination.TotalPagesPlaceholder;
 import com.github.syr0ws.fastinventory.common.placeholder.player.PlayerNamePlaceholder;
 import com.github.syr0ws.fastinventory.common.placeholder.player.PlayerUUIDPlaceholder;
@@ -17,7 +19,9 @@ public enum CommonPlaceholder {
     PLAYER_NAME(new PlayerNamePlaceholder()),
     PLAYER_UUID(new PlayerUUIDPlaceholder()),
     CURRENT_PAGE(new CurrentPagePlaceholder()),
-    TOTAL_PAGES(new TotalPagesPlaceholder());
+    TOTAL_PAGES(new TotalPagesPlaceholder()),
+    PREVIOUS_PAGE(new PreviousPagePlaceholder()),
+    NEXT_PAGE(new NextPagePlaceholder());
 
     private final Placeholder placeholder;
