+++ src/main/java/com/github/syr0ws/fastinventory/common/placeholder/pagination/TotalPagesPlaceholder.java
@@ -1,13 +1,9 @@
 package com.github.syr0ws.fastinventory.common.placeholder.pagination;
 
-import com.github.syr0ws.fastinventory.api.FastInventory;
-import com.github.syr0ws.fastinventory.api.pagination.Pagination;
 import com.github.syr0ws.fastinventory.api.pagination.PaginationModel;
-import com.github.syr0ws.fastinventory.api.placeholder.Placeholder;
 import com.github.syr0ws.fastinventory.api.util.Context;
-import com.github.syr0ws.fastinventory.common.CommonContextKey;
 
-public class TotalPagesPlaceholder implements Placeholder {
+public class TotalPagesPlaceholder extends PaginationPlaceholder {
 
     @Override
     public String getName() {
@@ -16,20 +12,7 @@ public class TotalPagesPlaceholder implements Placeholder {
 
     @Override
     public String getValue(Context context) {
-
-        FastInventory inventory = context.getData(CommonContextKey.INVENTORY.name(), FastInventory.class);
-        String paginationId = context.getData(CommonContextKey.PAGINATION_ID.name(), String.class);
-
-        Pagination<?> pagination = inventory.getPagination(paginationId)
-                .orElseThrow(() -> new IllegalStateException(String.format("No pagination with id %s found", paginationId)));
-
-        PaginationModel<?> model = pagination.getModel();
-
+        PaginationModel<?> model = super.getPaginationModel(context);
         return Integer.toString(model.countPages());
-    }
-
-    @Override
-    public boolean accept(Context context) {
-        return context.hasData(CommonContextKey.PAGINATION_ID.name());
     }
 }
